transparency_bundle/requirements.txt
psycopg2
records
agate
python-dotenv